requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
asgiref==3.5.0
    # via django
django==4.0.3
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
django-filter==21.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.1.0
    # via -r requirements.in
importlib-metadata==4.11.3
    # via markdown
markdown==3.3.6
    # via -r requirements.in
psycopg2-binary==2.9.3
    # via -r requirements.in
pyjwt==2.3.0
    # via djangorestframework-simplejwt
pytz==2022.1
    # via djangorestframework
sqlparse==0.4.2
    # via django
zipp==3.7.0
    # via importlib-metadata
